requirements.txt
wsgidav>=4.0.0
zeroconf>=0.47.0
pystray>=0.19.0
requests>=2.25.0
netifaces>=0.11.0
pyinstaller>=5.0
Pillow>=9.0.0
waitress>=2.1.2 # Added waitress for a production-grade WSGI server